Update Filament resource structure: campaign module
